+++ lib/public/components/Filters/common/FilteringModel.js
@@ -38,11 +38,15 @@ export class FilteringModel extends Observable {
     /**
      * Reset the filters
      *
+     * @param {boolean} notify call notify on self if true
      * @return {void}
      */
-    reset() {
+    reset(notify = false) {
         for (const model of this._filterModels) {
             model.reset();
+        }
+        if (notify) {
+            this.notify();
         }
     }
