+++ lib/public/components/Filters/common/filtersPanelPopover.js
@@ -45,7 +45,7 @@ const filtersToggleContent = (
             {
                 onclick: () => filteringModel.resetFiltering
                     ? filteringModel.resetFiltering()
-                    : filteringModel.reset(),
+                    : filteringModel.reset(true),
                 disabled: !filteringModel.isAnyFilterActive(),
             },
             'Reset all filters',
