+++ lib/public/views/DataPasses/ActiveColumns/dataPassesActiveColumns.js
@@ -35,7 +35,7 @@ export const dataPassesActiveColumns = {
         visible: true,
         sortable: true,
         format: (_, dataPass) => formatDataPassName(dataPass),
-        filter: ({ filteringModel }) => textFilter(
+        filter: (filteringModel) => textFilter(
             filteringModel.get('names'),
             { class: 'w-75 mt1', placeholder: 'e.g. LHC22a_apass1, ...' },
         ),
@@ -105,7 +105,7 @@ export const dataPassesActiveColumns = {
 
     nonPhysicsProductions: {
         name: 'Include nonphysics productions',
-        filter: ({ filteringModel }) => checkboxes(filteringModel.get('nonPhysics').selectionModel),
+        filter: (filteringModel) => checkboxes(filteringModel.get('nonPhysics').selectionModel),
         visible: false,
     },
 };
