+++ lib/public/views/DataPasses/PerLhcPeriodOverview/DataPassesPerLhcPeriodOverviewModel.js
@@ -10,34 +10,18 @@
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
  */
-import { FilteringModel } from '../../../components/Filters/common/FilteringModel.js';
-import { SelectionFilterModel } from '../../../components/Filters/common/filters/SelectionFilterModel.js';
-import { TextTokensFilterModel } from '../../../components/Filters/common/filters/TextTokensFilterModel.js';
-import { NON_PHYSICS_PRODUCTIONS_NAMES_WORDS, NonPhysicsProductionsNamesWords } from '../../../domain/enums/NonPhysicsProductionsNamesWords.js';
-import { OverviewPageModel } from '../../../models/OverviewModel.js';
+import { DataPassesOverviewModel } from '../DataPassesOverviewModel.js';
 import { buildUrl } from '/js/src/index.js';
 
 /**
  * Data Passes Per LHC Period overview model
  */
-export class DataPassesPerLhcPeriodOverviewModel extends OverviewPageModel {
+export class DataPassesPerLhcPeriodOverviewModel extends DataPassesOverviewModel {
     /**
      * Constructor
      */
     constructor() {
         super();
-        this._filteringModel = new FilteringModel({
-            names: new TextTokensFilterModel(),
-            nonPhysics: new SelectionFilterModel({
-                availableOptions: NON_PHYSICS_PRODUCTIONS_NAMES_WORDS.map((word) => ({ label: word.toUpperCase(), value: word })),
-            }),
-        });
-
-        this._filteringModel.visualChange$.bubbleTo(this);
-        this._filteringModel.observe(() => {
-            this._pagination.silentlySetCurrentPage(1);
-            this.load();
-        });
     }
 
     /**
@@ -52,27 +36,17 @@ export class DataPassesPerLhcPeriodOverviewModel extends OverviewPageModel {
      * @inheritdoc
      */
     getRootEndpoint() {
-        const nonPhysicsSelectionModel = this._filteringModel.get('nonPhysics').selectionModel;
+        const { filter, sort } = this.getBaseParams();
+        console.log(sort)
         const params = {
             filter: {
                 lhcPeriodIds: [this._lhcPeriodId],
-                names: this._filteringModel.get('names').normalized,
-                includeTest: nonPhysicsSelectionModel.selected.includes(NonPhysicsProductionsNamesWords.TEST),
-                includeDebug: nonPhysicsSelectionModel.selected.includes(NonPhysicsProductionsNamesWords.DEBUG),
+                ...filter,
             },
+            ...sort,
         };
 
         return buildUrl('/api/dataPasses', params);
-    }
-
-    /**
-     * Reset this model to its default
-     *
-     * @returns {void}
-     */
-    reset() {
-        this._filteringModel.reset();
-        super.reset();
     }
 
     /**
@@ -80,22 +54,5 @@ export class DataPassesPerLhcPeriodOverviewModel extends OverviewPageModel {
      */
     get lhcPeriodId() {
         return this._lhcPeriodId;
-    }
-
-    /**
-     * Return the filtering model
-     *
-     * @return {FilteringModel} the filtering model
-     */
-    get filteringModel() {
-        return this._filteringModel;
-    }
-
-    /**
-     * Checks if any filter value has been modified from their default (empty)
-     * @return {Boolean} If any filter is active
-     */
-    isAnyFilterActive() {
-        return this._filteringModel.isAnyFilterActive();
     }
 }
