+++ lib/public/views/DataPasses/PerLhcPeriodOverview/DataPassesPerLhcPeriodOverviewPage.js
@@ -50,7 +50,7 @@ export const DataPassesPerLhcPeriodOverviewPage = ({ dataPasses: { perLhcPeriodO
     return h('', {
         onremove: () => dataPassesPerLhcPeriodOverviewModel.reset(),
     }, [
-        h('.flex-row.header-container.pv2', filtersPanelPopover(dataPassesPerLhcPeriodOverviewModel, dataPassesActiveColumns)),
+        h('.flex-row.header-container.pv2', filtersPanelPopover(dataPassesPerLhcPeriodOverviewModel.filteringModel, dataPassesActiveColumns)),
         h('.w-100.flex-column', [
             table(
                 dataPassesPerLhcPeriodOverviewModel.items,
