+++ lib/public/views/DataPasses/PerSimulationPassOverview/DataPassesPerSimulationPassOverviewModel.js
@@ -10,31 +10,20 @@
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
  */
-import { SortModel } from '../../../components/common/table/SortModel.js';
-import { TextTokensFilterModel } from '../../../components/Filters/common/filters/TextTokensFilterModel.js';
-import { OverviewPageModel } from '../../../models/OverviewModel.js';
 import { buildUrl, RemoteData } from '/js/src/index.js';
 import { getRemoteData } from '../../../utilities/fetch/getRemoteData.js';
 import { ObservableData } from '../../../utilities/ObservableData.js';
+import { DataPassesOverviewModel } from '../DataPassesOverviewModel.js';
 
 /**
  * Data Passes Per Simulation Pass overview model
  */
-export class DataPassesPerSimulationPassOverviewModel extends OverviewPageModel {
+export class DataPassesPerSimulationPassOverviewModel extends DataPassesOverviewModel {
     /**
      * Constructor
      */
     constructor() {
         super();
-        this._sortModel = new SortModel();
-        this._sortModel.observe(() => {
-            this._pagination.silentlySetCurrentPage(1);
-            this.load();
-        });
-        this._sortModel.visualChange$.bubbleTo(this);
-
-        this._nameFilterModel = new TextTokensFilterModel();
-        this._registerFilter(this._nameFilterModel);
         this._simulationPass = new ObservableData(RemoteData.notAsked());
         this._simulationPass.bubbleTo(this);
     }
@@ -65,17 +54,14 @@ export class DataPassesPerSimulationPassOverviewModel extends OverviewPageModel 
      * @inheritdoc
      */
     getRootEndpoint() {
+        const { filter, sort } = this.getBaseParams();
         const params = {
             filter: {
-                names: this._nameFilterModel.normalized,
                 simulationPassIds: [this._simulationPassId],
+                ...filter,
             },
+            ...sort,
         };
-
-        const { appliedOn: sortOn, appliedDirection: sortDirection } = this._sortModel;
-        if (sortOn && sortDirection) {
-            params[`sort[${sortOn}]`] = sortDirection;
-        }
 
         return buildUrl('/api/dataPasses', params);
     }
@@ -89,58 +75,9 @@ export class DataPassesPerSimulationPassOverviewModel extends OverviewPageModel 
     }
 
     /**
-     * Reset this model to its default
-     *
-     * @returns {void}
-     */
-    reset() {
-        this._nameFilterModel.reset();
-        super.reset();
-    }
-
-    /**
      * Get current simulation pass which data passes are fetched
      */
     get simulationPass() {
         return this._simulationPass.getCurrent();
-    }
-
-    /**
-     * Returns the model handling the overview page table sort
-     *
-     * @return {SortModel} the sort model
-     */
-    get sortModel() {
-        return this._sortModel;
-    }
-
-    /**
-     * Returns data pass name filter model
-     * @return {TextTokensFilterModel} data pass name filter model
-     */
-    get nameFilterModel() {
-        return this._nameFilterModel;
-    }
-
-    /**
-     * Register a new filter model
-     * @param {FilterModel} filterModel the filter model to register
-     * @return {void}
-     * @private
-     */
-    _registerFilter(filterModel) {
-        filterModel.visualChange$.bubbleTo(this);
-        filterModel.observe(() => {
-            this._pagination.silentlySetCurrentPage(1);
-            this.load();
-        });
-    }
-
-    /**
-     * States whether any filter is active
-     * @return {boolean} true if any filter is active
-     */
-    isAnyFilterActive() {
-        return !this._nameFilterModel.isEmpty();
     }
 }
