+++ lib/public/views/DataPasses/PerSimulationPassOverview/DataPassesPerSimulationPassOverviewPage.js
@@ -59,7 +59,7 @@ export const DataPassesPerSimulationPassOverviewPage = ({ dataPasses: {
         onremove: () => dataPassesPerSimulationPassOverviewModel.reset(),
     }, [
         h('.flex-row.items-center.g2', [
-            filtersPanelPopover(dataPassesPerSimulationPassOverviewModel, dataPassesActiveColumns),
+            filtersPanelPopover(dataPassesPerSimulationPassOverviewModel.filteringModel, dataPassesActiveColumns),
             h(
                 '.flex-row.g1.items-center',
                 simulationPass.match({
